feat: change dynamic block behavior
Change dynamic blocks to have a many-to-many relation with agents
since application and user scopes requires the removal of agentId reference
from dynamic block table.
Add a new interface to add dynamic blocks just like a prompt template, the dynamic block
is automatically created when detected as a double curly bracket within the system prompt
of an agent by the PromptTemplate component. It now autofills the system prompt when something
is wrapped with double curly brackets